backend/pyproject.toml
[tool.poetry]
name = "api-test-gen-backend"
version = "0.1.0"
description = "API Test Generation Backend"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"
pgvector = "^0.2.4"
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
prance = "^0.22.0"
openapi-spec-validator = "^0.7.1"
langchain = "^0.0.350"
openai = "^1.3.5"
schemathesis = "^3.19.0"
requests = "^2.31.0"
httpx = "^0.25.0"
pyyaml = "^6.0.0"
locust = "^2.17.0"
cryptography = "^41.0.7"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
celery = {extras = ["redis"], version = "^5.3.4"}
redis = "^5.0.1"
prometheus-client = "^0.19.0"
sentry-sdk = {extras = ["fastapi"], version = "^1.38.0"}
faker = "^20.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

